move the `log` func to core.utils
