+++ tdrs-backend/tdpservice/core/utils.py
@@ -1,5 +1,11 @@
 """Core utility classes and functions."""
 
+import logging
+from django.contrib.admin.models import LogEntry, ContentType, CHANGE
+
+
+logger = logging.getLogger()
+
 
 class ReadOnlyAdminMixin:
     """Mixin to enforce read-only models in Django Admin.
@@ -20,3 +26,34 @@ class ReadOnlyAdminMixin:
     def has_delete_permission(self, request, obj=None):
         """Deny all delete permissions."""
         return False
+
+
+def log(msg, logger_context=None, level='info'):
+    """Create a log in the terminal and django admin console, for email tasks."""
+    from tdpservice.users.models import User
+
+    log_func = logger.info
+
+    match level:
+        case 'info':
+            log_func = logger.info
+        case 'warn':
+            log_func = logger.warn
+        case 'error':
+            log_func = logger.error
+        case 'critical':
+            log_func = logger.critical
+        case 'exception':
+            log_func = logger.exception
+
+    log_func(msg)
+
+    if logger_context:
+        LogEntry.objects.log_action(
+            user_id=logger_context['user_id'],
+            change_message=msg,
+            action_flag=CHANGE,
+            content_type_id=ContentType.objects.get_for_model(User).pk,
+            object_id=logger_context['object_id'],
+            object_repr=logger_context['object_repr']
+        )
